function-app/requirements.txt
azure-functions>=1.18.0
azure-storage-blob>=12.19.0
azure-identity>=1.15.0
azure-keyvault-secrets>=4.7.0
pyodbc>=5.0.1
lxml>=5.1.0
requests>=2.31.0
